door styles size code added, renamed jadeDoors, renamed VanaraTree,
